backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
aiosqlite